Move progress module to common scope
