+++ tpcgen-cli/src/lib.rs
@@ -1,5 +1,6 @@
 //! This library contains both the TPCH and TPCDS command line clients.
 mod logging;
 mod parquet;
+pub mod progress;
 pub mod tpcds_cli;
 pub mod tpch_cli;
