+++ tpcgen-cli/src/tpch_cli.rs
@@ -8,11 +8,11 @@ pub mod generate;
 pub mod output_plan;
 pub mod parquet;
 pub mod plan;
-pub mod progress;
 pub mod runner;
 pub mod statistics;
 pub mod tbl;
 
+pub use crate::progress;
 pub use cli::Cli;
 pub use generator::{
     Compression, GeneratorConfig, OutputFormat, Table, TpchGenerator, TpchGeneratorBuilder,
