+++ tpcgen-cli/src/tpch_cli/parquet.rs
@@ -1,6 +1,6 @@
 //! TPCH Parquet output format.
 
-use crate::tpch_cli::progress::ProgressTracker;
+use crate::progress::ProgressTracker;
 use arrow::record_batch::RecordBatchReader;
 use parquet::basic::Compression;
 use std::io;
